refact : auth user 통합
